receiver_creator: add dynamic configuration values (#235)
This adds the ability to configure a receiver based on values from the
observer. This might be from labels or some other metadata the user sets. This
includes setting the port if the port is not able to be discovered through
standard means.

For example, in k8s it is optional for a pod to define the ports containers are
listening on. Users may need to set those ports inside the config. This is
possible with dynamic configuration.

Will add docs to readme after other PR is merged.

Example:

```yaml
receivers:
  receiver_creator:
    redis:
      rule: type.pod && labels["app"] == "redis"
      config:
        endpoint: "`endpoint`:8382"
```

The text between backticks (in this case `endpoint`) is evaluated using expr
(same type of expressions as with discovery rules).