Add common signalfx host id extraction (#1100)
This adds a shared way to extract AWSUniqueId and gcp_id dimensions. It will be
used by both sapmexporter and signalfxexporter. signalfxexporter cannot
currently use the method because it's not on pdata and it seemed silly to
unconvert metrics data back to pdata. I've added a note to signalfxexporter to
use this common implementation once it switches to pdata.