exporter/signalfx: Remove send_compatible_metrics option (#2267)
* exporter/signalfx: Remove send_compatible_metrics option

This is a **breaking change**. Remove `send_compatible_metrics` options. Instead
just use the `translation_rules` field to determine whether or not compatible
metrics are to be sent. If `translation_rules` is not specified, default rules
will be applied and compatible metrics will be sent. Compatible metrics can be
turned off by setting it to `[]`.

* Update comment in custom unmarshaler