exporter/signalfx: Handle character limits on metric names and dimensions (#3328)
The `signalfx` exporter currently does not account for character limit sizes of metric names or dimensions. As a result, telemetry that does not meet character limits specified [here](https://dev.splunk.com/observability/docs/datamodel/ingest#Criteria-for-metric-and-dimension-names-and-values) are sent to the backend. The backend returns 200s on receipt of such data. Hence, some sanitization needs to happen at the exporter. Otherwise datapoints will silently be dropped by the backend. This update does the following - 

- drops datapoints when metric name is invalid (too long)
- drops dimensions on datapoints when dimension key or value is invalid (too long)