[exporter/datadog]: add collector version tag to internal health metrics (#3394)
This PR adds the opentelemetry-collector `version` as a tag on metrics sent via the datadogexporter to datadog about the health and uptime of the underlying opentelemetry-collector `otel.datadog_exporter.(metrics|traces).running`. This tag will be used to help debugging, troubleshooting, and understanding which versions of the collector are deployed in a user's environment.