[pkg/telemetryquerylanguage] Create reusable StandardGetSetter (#12597)
In order to facilitate moving functions to their own packages as part of #11751, we need a place to put testGetSetter so it could be reusable. That need cascaded into creating a basic GetSetter that ends up being useful everywhere. Instead of creating a new testLiteral, the existing Literal struct was exposed.