[pkg/stanza/fileconsumer] Process batches consecutively, rather than one per poll interval (#12910)
[pkg/stanza/fileconsumer] Process batches consecutively, rather than one per interval

Prior to this change, batches were processed at a rate of one per poll interval.
This was never considered the desired behavior, but it fit the codebase more
easily. This change pulls the batching logic into a higher level function, and
consecutively dispatches batches to a lower level 'consume' function.