[pkg/stanza] Fix bug where original severity text was not preserved (#13265)
When converting from entry.Entry to plog.Logs, the severity text
was not copied over. Instead, it was automatically assigned based
on the severity number.

This change does two things:
1. If the severity text is not empty, it is preserved
2. If the severity text is empty, it is assigned based on the severity
number. However, the values assigned were also incorrect, in that they
did not match the case defined in the data model. Previously, a
capitalized style (eg "Info") was used. Now, an all caps style is used.
It is rare that severity number is specified while severity text is not,
but the data model explicitly calls for this handling, so it is implemented
as such.