[receiver/vcenter] Fix incorrect KBy and MBy units (#14108)
The vcenter receiver indicates that it uses kilobytes (kBy), but instead the units are actually kibibytes (KiBy). The vSphere API docs claims they are in kilobytes, but documentation on the data counters units indicates that they use kilobytes to mean 1024 bytes, which in UCUM is actually kibibytes (KiBy).

Similar logic applies to the MBy units. VMware, in general, seems to prefer returning in power of 2 units instead of power of 10 units.