[receiver/zipkinreceiver] Fix Zipkin V1 receiver creating invalid timestamps (#16062)
I think this issue was introduced in b2add3cd01c1a24c0cc85503b302d91493c95794.

I [noticed a memory leak][1] in DataDog exporter but after a lot of
debugging, it turned out that the OpenTelemetry gateway was receiving
invalid timestamps:

```
ScopeSpans #0
ScopeSpans SchemaURL:
InstrumentationScope
Span #0
    Trace ID       : 0000000000000000c20a2b82c179228a
    Parent ID      :
    ID             : 7c3415ed370f1777
    Name           : [redacted]
    Kind           : SPAN_KIND_SERVER
    Start time     : 2200-11-16 22:32:41.14035456 +0000 UTC
    End time       : 2200-11-16 22:33:02.68735456 +0000 UTC
```

The year is almost 100 years in the future.

Before we send a span to the gateway, it goes through an OpenTelemetry
collector sidecar. We use Zipkin V1 endpoint to send the spans to the
sidecar.

In the same pod, we're also using Zipkin V2 which did not have any
issues.

The problem itself: zipkin uses microseconds. This has to be multipled
with 1e3 to get nanoseconds, not with 1e6.

[1]: https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/15720