[exporter/prometheus] Exemplars for monotonic sums (#18307)
This will add metrics exemplars to the Prometheus exporter for monotonic sum metrics (i.e. counters in Prometheus). Note that this could possibly be a breaking change if users are producing monotonic sums with exemplars, have enabled open metrics export format in this exporter and exemplar ingest in Prometheus, but have not enabled the pkg.translator.prometheus.NormalizeName feature gate and their metric names do not already follow the convention required by OpenMetrics, i.e. the name ends in _total. In the breaking cases, the exporter reports the type of the metric as unknown and then Prometheus appears to drop the metric on ingest due to the presence of the exemplar data.