[receiver/dockerstats] support pids_stats metrics (#21111)
pids_stats metrics were introduced in docker engine API v1.23 and that's the reason why docker's API default version has been bumped to 1.23. However, the minimum supported version has not been updated because the metrics are optional and reported only when pids cgroup is supported and the kernel is >= 4.3. On top of that, pids_stats may be reported even if the API request is performed using previous versions.