[connector/*] generate status from metadata.yaml (#22800)
This PR adds the ability to generate status from metadata.yaml for connectors. The format for connectors differs in that stability is per connection rather than per signal as it is for other components. In addition to the code to generate status, this uses the count connector as a demonstration of how this will work in the real world.