[chore][processor/k8sattributes] Move generic testing logic to separate package (#23228)
The k8sattributes processor has an e2e test with logic that is not specific to the processor itself. This testing logic will be helpful for other k8s e2e tests, so this moves that logic into its own package. Also added a README to explain current abilities and usage.

Note: This PR is not as big as it looks. There is no functionality change, it's simply moving functions into separate files and a new package. The test change itself is simply to reference the functions and variables in their new package, rather than its own file.

Feedback on file names and which functions should go in which files is very welcome, the point of this is to make the helpers easy to find and use.