[internal/filter] Move filtermetric config to filterconfig (#23142)
**Description:** 
Moves filtermetric's config to filterconfig to facilitate a better
bridge to OTTL. Rename to `Metric*` to not conflict with existing config
structs.

This PR attempts to be as small as possible - only the config and its
functions are moved/changed. How components use those configs does not
change. This PR does not attempt to combine the metric configuration
with the general configuration.

**Link to tracking Issue:** 
Check
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/23141
to see how this helps the filterottl bridge.