[extension/headers_setter] Export ActionValue (#24187)
This PR exports the `ActionValue` type in headers_setter's config, for
these reasons:
* Configuration types are generally always exported. See `HeaderConfig`
in the headers_setter's config for example.
* It makes it easier to integrate Collector features in applications
such as Grafana Agent. Making `ActionValue` external would allow us to
create `Convert()` functions which return a type `ActionValue`, in order
to convert Grafana Agent config to Collector config.

There is no changelog entry because this is not a customer-facing
change.