[processor/resourcedetection] Add `host.arch` to system detector (#22940)
**Description:** 

Add `host.arch` to system detector. It is detected via `runtime.GOARCH`.

This way of detecting may be wrong if running under an emulator or
binary translator; the only common case I know of this happening in
practice is on ARM64 macOS when using Rosetta. There is a way to detect
this (see DataDog/gohai/pull/77) but since we provide arm64 builds for
macOS, I think this is a non-issue

**Link to tracking Issue:** Fixes #22939

**Testing:** Added unit tests. Tested on my Linux AMD64 laptop

**Documentation:** Documented new attribute on README