[chore] Add "Code Owners" field to new component template (#23780)
#20868 represents a push to ensure that all component code owners are
members of the OpenTelemetry organization.

This has direct implications for our notion of "vendor-specific
components", which are those that interact specifically with a
particular vendor, and are contributed and maintained by a
representative of a vendor. We require that the vendor representative
must maintain the component, which essentially means they must be a code
owner. Therefore, I am proposing that the representative must be a
member of the org before the component may be accepted.