[extension/jaegerremotesampling] gRPC remote mode propagates configured HTTP headers (#24414)
**Description:** This could be described as a fix or an improvement to
the [jaegerremotesampling
extension](https://github.com/open-telemetry/opentelemetry-collector-contrib/tree/main/extension/jaegerremotesampling).

Status quo: For usages in which a `remote` source is specified using
collector-wide standard
[configgrpc.GRPCClientSettings](https://github.com/open-telemetry/opentelemetry-collector/blob/b5e511ce31f22fd3d4817236792245fe1bd88ef8/config/configgrpc/configgrpc.go#L54),
the given HTTP headers are not actually set on outbound calls to the
destination `grpcstore.SamplingManager` endpoint.

After this PR: Outbound calls will add any HTTP headers specified in the
gRPC client settings for the remote source. This will mean that drop-in
extension usage will support use cases in which header additions are
necessary for remote interactions. I took an approach that I observed in
several other exporters/extensions: "enhancing" the gRPC context.

**Link to tracking Issue:** N/A

**Testing:** Existing extension integration tests have been updated to
perform a (previously not performed) client-like HTTP call to the
extension's running gRPC server, and then to verify that an observed
call to a gRPC remote includes configured HTTP header additions (as gRPC
metadata).

**Documentation:** N/A: I assumed the behavior that this PR now
implements, because the gRPC client settings config is so "standard"
throughout the opentelemetry-collector repo (and other extensions in
this contrib repo).

---------

Co-authored-by: Antoine Toulme <antoine@toulme.name>