[processor/transform] Add extract_count_metric OTTL function (#24897)
**Description:**
Added a function for extracting the count from a Histogram,
ExponentialHistogram or Summary as an individual metric. The added
function also works for Summaries, so we can later deprecate the
Summary-specific `summary_count_val_to_sum` function.

This is very similar to #24368, which added a function for extracting
the sum in the same fashion. The only significant difference is that
count values are required by the spec. See that PR for implementation
discussion.

**Link to tracking Issue:** #22853 

**Testing:**
Added unit tests both for the function itself and OTTL statements
involving it.

**Documentation:**
Added a section in the README.