[exporter/awsxray] Feature Gate to Allow fully random Trace IDs in XRay Exporter  (#26041)
Currently, the AWS XRay Exporter will drop segments/traces with trace
IDs where the first 32 bits, when converted to UNIX Epoch time, are not
within the past 28 days. This change is to add an Alpha feature gate
`exporter.awsxray.skiptimestampvalidation`, which is disabled by
default. If enabled, the timestamp restriction is removed so that users
do not need to use the AWS XRay ID Generator.

- Updated unit test in `config_test.go` 
- Added unit tests in `factory_test.go` and `segment_test.go`
- All other unit tests related to trace ID are passing.