[chore][pkg/stanza] Simplify tokenize package (#26040)
This PR contains several simplifications within and relating to the new
tokenization package:
- `tokenize.Multiline` now returns a `bufio.splitFunc` without concern
for a flush func. A flush func is applied later if necessary.
- Remove `tokenize.Splitter`. `SplitterConfig` can now generate a
`bufio.SplitFunc` directly.
- Unexport `tokenize.Flusher`. `FlusherConfig` can now wrap a
`bufio.SplitFunc` directly. (It does still create a struct to maintain
state, but then wraps the split func with a receiver that interacts with
this state.)
- `fileconsumer`'s internal `splitter` package was also simplified. It
may make sense to merge this into `tokenize` in a future PR, but this PR
is large enough.