[pkg/stanza] Extract trim func from split package (#26536)
Follows
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/26241

Previously, split funcs were responsible for applying trim funcs. This
PR increases composability by applying trim funcs as a wrapper around
split funcs.

One nuance that was surfaced here is that the newline split func was not
handling the case where a line starts with a newline. When this happens,
we need to tell the scanner to advance, but we still want to return a
`""` token, rather than nil. This is covered by existing tests, but
previously it was "fixed" by the trim func which would return an empty
slice when the token was nil. Now, the newline split func will
explicitly handle this case, while the trim func will return the
original value if it is nil or empty.