Warn instead of failing when renaming metrics using metric_relabel_configs (#25888)
**Description:**

Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/5001

When that issue was filed, we dropped unknown-typed metrics. Now that we
convert them to gauges, users may want to use metric_relabel_configs to
rename metrics, even if they lose metadata. In the future, we have other
enhancements planned to better-support unknown-typed metrics.

Change the error to a warning. It is still likely not a best practice to
use relabel configs for this purpose, but for users copy-pasting
prometheus configs, it is best not to fail if we can help it.