[headerssetter] Remove "Limitations" section from README. (#27465)
The docs for the `headers_setter` extension state that the
`from_context` config option does not work if the batch processor is
used. I believe this comment is out of date, because it seems that the
[PR](https://github.com/open-telemetry/opentelemetry-collector/pull/7578)
linked to the
[issue](https://github.com/open-telemetry/opentelemetry-collector/issues/4544)
has been merged. Note that I haven't tested that this actually works
with the header setter extension. Please let me know if you think
testing is necessary.

Co-authored-by: Juraci Paixão Kröhling <juraci@kroehling.de>