[pkg/ottl] Add ConditionSequence struct (#29339)
**Description:** 

Adds a new `ConditionSequence` struct to help handle processing a list
of Conditions. The primary reason to use a `ConditionSequence` is to let
the struct handle errors. Since that is its defining purpose, I opted to
make `ErrorMode` a required argument in the "constructor" instead of an
Option.

I also update `internal/filterottl` to use this struct instead of
`Statements`. This is a non-breaking change.

If we like this pattern, I will do a breaking change to replace
`Statements` with a similar `StatementSequence` struct in a future PR .

See these structs implemented in components here:
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/29294

**Link to tracking Issue:** <Issue number if applicable>

Related to
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/13545

**Testing:**

Added new tests

**Documentation:** <Describe the documentation added.>
Added godoc comments

---------

Co-authored-by: Evan Bradley <11745660+evan-bradley@users.noreply.github.com>