[chore][cmd/otelcontribcol] Re-enable exporters and receivers tests on Windows (#29532)
**Description:**
Last part to re-enable all Windows tests disabled due to issue #11451.

Notes:

* The default test timeout is not enough for the `cmd` GROUP on CI and
it was increased to 1200s just for this group on Windows, the default
stays at 300s (which was also enough for my local run).
* The `fileexporter` life cycle test indicates an issue that shouldn't
affect most usages of the collector: if the same configuration is used
to export multiple signals it opens one file handle for each signal but
just closes the one of them at shutdown (the helper only executes the
shutdown for the one signal). This is an issue for the test because
during cleanup the testing attempts to delete the test folder and there
are still handles open to the file that lives inside it. This shouldn't
be a problem for collector users. That said the exporter should close
all file handles on shutdown. I will take look at what can be done to
correct this behavior.
* Fixed a minor typo on the receivers test that affected many lines.

**Link to tracking Issue:**
Fix #28679 - last part.

**Testing:**
Local run and GH CI on my fork with `Run Windows` label.

**Documentation:**
N/A