[pkg/ottl] Add private interface for path and key (#29704)
**Description:**
This is the first PR working towards
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/22744.

It adds 2 new interfaces, `path` and `key`, that will be given to
contexts to use when interpreting the paths/keys passed from the
grammar. Eventually we'll be able to use the `isComplete` functions
during parsing to ensure that all parts of the path were used by the
context.

By the end of that issue we'll have removed all public API's of OTTL's
grammar, which is going to cause massive breaking changes to all our
contexts and huge PRs. To help decrease the amount of changes per PR
these new interfaces are not yet exported and not yet implemented. To
see them partially in use checkout
https://github.com/open-telemetry/opentelemetry-collector-contrib/pull/29703.

**Link to tracking Issue:**

Related to
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/22744

**Testing:**

Added unit tests

**Documentation:** <Describe the documentation added.>

I'll add godoc comments once these things are made public

---------

Co-authored-by: Evan Bradley <11745660+evan-bradley@users.noreply.github.com>