[receiver/snowflake] Fix storage metric reporting bug (#29775)
**Description:** Identified and fixed an issue with the function
responsible for querying storage metrics in `client.go`. The issue was
that `sql.Rows.Scan()` was receiving a `float64` instead of an `int64`
as previously thought and so trying to copy the column values into the
provided variable pointers was causing a type error. To fix this the
type of the variables passed into `sql.Rows.Scan()` was changed to
`float64` and then recast as `int64` when inserting into the metric
interface. This made more sense than potentially breaking existing
pipelines by changing the metric type to `double` especially since the
values for these metrics is always an integer (even if the database
itself stores them as a float).

Fixes #29750

**Testing:** Modified tests to expect float values from the mock
database