[extension/filestorage] enable feature gate `extension.filestorage.replaceUnsafeCharacters` (#30755)
**Description:**

Moves the feature gate `extension.filestorage.replaceUnsafeCharacters`
to Beta stage, which means it is now enabled by default. This is
according to the schedule for this feature gate described in the File
Storage extension's
[README](https://github.com/open-telemetry/opentelemetry-collector-contrib/blob/v0.92.0/extension/storage/filestorage/README.md#extensionfilestoragereplaceunsafecharacters).

**Link to tracking Issue:**

-
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/3148

**Testing:**

Updated tests to not require enabling of the feature gate.

**Documentation:**

Updated version in which feature gate was moved to Beta.