[chore][processor/cumulativetodeltaprocessor] Enable goleak check (#30804)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
This enables the `goleak` check on all tests in this package, to ensure
the processor isn't leaking goroutines. The only update here is in
testing, as one test was missing a necessary `shutdown` call.

**Link to tracking Issue:** <Issue number if applicable>
#30438

**Testing:** <Describe what testing was performed and which tests were
added.>
All tests are passing, including added `goleak` check.