[chore] Enable goleak for tests failing on opencensus-go (#30457)
**Description:** 
This enables `goleak` checks for all packages whose only failure is due
to [a known
issue](https://github.com/census-instrumentation/opencensus-go/issues/1191)
within the `opencensus-go` package. I believe it's safe to ignore for
two main reasons:
1. All references that I'm aware of are indirect, the contrib repo isn't
directly using the stat's worker functionality. The goroutine is started
in the dependency's `init()` so it's not something that we're causing.
2. We're actively moving away from the opencensus dependency, and hope
to remove it entirely soon. Relevant issue:
https://github.com/open-telemetry/opentelemetry-collector/issues/816.
Once this work is complete the ignore will no longer be necessary and we
can safely remove it.

The only changes in this PR are test related. The same file is being
copied into different packages, and then we run `make gotidy`. No
functionality has been changed.

**Link to tracking Issue:** #30438