[chore] Do more steps in parallel in build-and-test workflow (#30891)
The build-and-test workflow takes a long time. Part of the issue is that
we spend a lot of time waiting for long jobs, such as integration tests,
to finish before moving onto more long jobs.

This PR attempts to parallelize more steps by removing some existing
ordering we've restricted ourselves to. Specifically this moves the
cross-compile chain to start in parallel with the unit tests, lint, and
integration tests. My reasoning is that if a unit tests or lint is bad
it will still get caught and we don't need unit tests to pass to attempt
to compile the code. I have update the publish-check step to still
depend on all these steps.

Related to
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/30880