[chore][pkg/stanza/operator/transformer/recombine] Enable goleak check (#31095)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
This enables `goleak` checks on the
`pkg/stanza/operator/transformer/recombine` package, to help ensure no
goroutines are leaking. This is a test only change, adding a missing
`Stop` call, as well as stopping a running goroutine in a test.

**Link to tracking Issue:** <Issue number if applicable>
#30438

**Testing:** <Describe what testing was performed and which tests were
added.>
All existing tests are passing, as well as added `goleak` check.

---------

Co-authored-by: Daniel Jaglowski <jaglows3@gmail.com>