[telemetrygen] Fix case where root CAs should inherit from host environment (#31250)
**Description:** 

Adding TLS configuration to telemetrygen made it so that providing root
CA information is mandatory. This is for the case when you are
generating CAs and using them to sign certificates. In the case where
the CA that signed the certificate is a known trusted CA, we shouldn't
have to provide its CA certificate. This PR fixes this problem by
allowing the code to pull its CA pool from the host environment when CAs
are not explicitly supplied.

**Link to tracking Issue:**
[github.com/open-telemetry/opentelemetry-collector-contrib/issues/31191](https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/31191)

**Testing:** Generate metrics, logs, traces against http and grpc
endpoints with legitimate certificates

**Documentation:** <Describe the documentation added.>