[chore][exporter/logzioexporter] Enable goleak check (#31151)
**Description:** <Describe what has changed.>
<!--Ex. Fixing a bug - Describe the bug and how this fixes the issue.
Ex. Adding a feature - Explain what this achieves.-->
This enables `goleak` checks in the Logz.io exporter to help ensure no
goroutines are being leaked. This is a test only change as the only
thing necessary was to add a missing close call on a server in a test.

**Link to tracking Issue:** <Issue number if applicable>
#30438

**Testing:** <Describe what testing was performed and which tests were
added.>
All existing tests are passing, as well as added `goleak` check.