Prometheus Receiver: Unregister self-observability metrics on Shutdown (#32202)
**Description:**

Prometheus libraries will currently fail to initialize if they are
called more than once for the same component name. This is because it
attempts to register the same self-observability metrics, which isn't
allowed by the Prometheus client. To work around this, Unregister all
self-observability metrics during Shutdown.

This updates the prometheus dependency to an unreleased version to pull
in https://github.com/prometheus/prometheus/pull/13897 and
https://github.com/prometheus/prometheus/pull/13896.

**Link to tracking Issue:** <Issue number if applicable>

Fixes
https://github.com/open-telemetry/opentelemetry-collector-contrib/issues/32123

**Testing:**

Unit test added.

**Documentation:**

N/A

cc @cpheps @djaglowski